requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
Django==3.1.3
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phone-field==1.8.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
PyJWT==2.0.0
pylint==2.6.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
